Don't analyze HIE files when hs file doesn't exist
There doesn't seem to be anything that cleans up corresponding .hie
files when hs files are removed. When weeder analyzes them it
(correctly) discovers all code as unused. This can be particularly
troublesome on CI where build artifacts (e.g. .stack-work) are heavily
cached. Removing a file will result in a Weeder error that is
non-obvious and requires an expensive operation (busting the cache) to
correct.

Fixes #6 and #48.