Locate .hs files differently
The HIE files capture source file paths as relative to the compilation
root, which may be within a sub-directory in the case of a multi-package
project. There seems to be no stable way to know which sub-directory to
look in, so this approach looks into them all.

It does this by first locating all .hs files and then looking for any
element with the candidate file as a suffix. This of course suffers from
false positives, but is hopefully acceptable. Projects which use package
names in their module namespaces should be fine, but some notable cases
of false-positive would be test/Spec.hs or executables/Main.hs.