Don't analyze HIE files when hs file doesn't exist (#82)
There doesn't seem to be anything that cleans up corresponding .hie
files when hs files are removed. When weeder analyzes them it
(correctly) discovers all code as unused. This can be particularly
troublesome on CI where build artifacts (e.g. .stack-work) are heavily
cached. Removing a file will result in a Weeder error that is
non-obvious and requires an expensive operation (busting the cache) to
correct.

Fixes #6 and #48.

The HIE files capture source file paths as relative to the compilation
root, which may be within a sub-directory in the case of a multi-package
project. There seems to be no stable way to know which sub-directory to
look in, so this approach looks into them all.

It does this by first locating all .hs files and then looking for any
element with the candidate file as a suffix. This of course suffers from
false positives, but is hopefully acceptable. Projects which use package
names in their module namespaces should be fine, but some notable cases
of false-positive would be test/Spec.hs or executables/Main.hs.

Co-authored-by: patrick brisbin <pbrisbin@gmail.com>