Add --no-exit-code option
In freckle/weeder-action, we don't want weeder to exit non-zero when
weeds are found because that would fail the Job. We want to capture and
process things ourselves.

Without this option, we had to resort to `|| true`, which means we don't
behave correctly when there are non-weed problems (failure to read
config, for example). These cases just appear successful. This is
particularly bad for new projects which don't have a `weeder.dhall` yet.
The team is never alerted to add it because the weeder-action just
succeeds, even though that should be a fatal error.

In adding the option, I copied HLint exactly in the name of the flag and
the help message. I find the language a little stilted, to be honest, so
I'm totally open to making it whatever you like; I just through being
consistent with an existing tool could be valuable.