Improving performance (#137)
* outputableDeclarations optimisation

We now only match on declarations that can appear in the output (i.e.
with a span) when evaluating `roots` and `dead` in `runWeeder`,
instead of searching through all of them.

* initialGraph optimisation

`addAllDeclarations` turned out to largely be an unnecessary step:
`define` and `addDependency` already created vertices for the
involved declarations on their own.

This skips that step when unused-types = false, and otherwise
minimises the number of added vertices.

* Compile regex in Config.hs

GHC does not optimise repeated calls to `=~`, so we were
recompiling the regex for every single declaration we matched on.
This took up 15% of runtime and 40% of memory usage.

This compiles regular expressions in advance, immediately after we
parse the config file. Besides performance, this has the additional
benefit of making `runWeeder` a total function: regex parse
failures are now emitted alongside TOML errors.

* Parallel analysis

We now run most of our analysis on each `HieFile` in parallel.

Doing this in batches of 100 turned out to not have any performance
difference from doing this with 1 `HieFile` at a time.

Given enough cores, this can almost double Weeder's performance.

* getHieFiles as a stream

This uses lazy IO via `getChanContents` to begin analysis in
`runWeeder` before `[HieFile]` is fully read.

This neither benefits nor harms performance when Weeder is given
only one core, but improves performance by about the runtime of
`getHieFiles` when given two or more cores. On my machine this is
15%, but this may vary depending on disk speed.

As following uses of type class instances requires information from
all HIE files, it is now done in a separate stage via
`analyseEvidenceUses`, applied only after the initial analysis is
fully evaluated.

* Fix throwing exit code 2 in getHieFiles

Switches `forkIO` to `async` in order to rethrow exceptions
encountered by `getHieFiles`.

* Replace type families with lists

* Set capabilities with -j and -N

* Update comments in Weeder.hs

---------

Co-authored-by: Ollie Charles <ollie@ocharles.org.uk>